function_source/requirements.txt
pulumi-gcp
psycopg2-binary>=2.8
